MainAppTest
yml:add rate and base currency. New quotaion param timestamp, default_date=now date
DTOResponseCurrency change double on float

+++ src/main/java/com/undefined/undefined/feign/dto/DTOResponseCurrency.java
@@ -13,7 +13,7 @@ public class DTOResponseCurrency {
     // Необязательно
     private long timestamp; // TODO переделать ли в норм дату???
     private String base;
-    private Map<String, Double> rates; // Котировки представляются в виде мапа
+    private Map<String, Float> rates; // Котировки представляются в виде мапа
 
     @Override
     public String toString() {
